+++ src/tabs/consumer.ts
@@ -16,16 +16,16 @@ export class TabsConsumer extends Consumer implements tabs.TabsConsumer {
     return protectedProvidersWeakMap.get(this)!.getTab(tabId);
   }
 
+  updateTab(tabId: string, updatedTab: tabs.TabUpdate): void {
+    return protectedProvidersWeakMap.get(this)!.updateTab(tabId, updatedTab);
+  }
+
   navToTab(tabId: string, options: tabs.NavToTabOptions = { silent: false }) {
     return protectedProvidersWeakMap.get(this)!.navToTab(tabId, options);
   }
 
   nav() {
     return protectedProvidersWeakMap.get(this)!.nav();
-  }
-
-  dispatchUrlInTab(tabId: string, url: string) {
-    return protectedProvidersWeakMap.get(this)!.dispatchUrlInTab(tabId, url);
   }
 
   getTabWebContentsState(tabId: string) {